requirements.txt
google-generativeai
langchain
beautifulsoup4
spacy
pandas
scipy
langchain_text_splitters
requests